fix code in buttonspost component
